Much better logic for handling retry time interval for feature matching tasks: - add an isImageDisplayed property to the feature matching record - property set to YES currently if image is in display list (by previous checkin) - if a match fails, there are now 5 possible cases to set the time interval to elapse before trying again
1) 1.5 seconds - invalid homography, image displayed
2) 3.0 seconds - invalid match (no feature correspondence, etc.), image displayed
3) 6.0 seconds - invalid homography, image not displayed
4) 6.0 seconds - invalid match (no feature correspondence, etc.), image not displayed
5) 0.0 seconds - error reading camera frame features from memory buffer (will re-queue next possible attempt)

This should have a much better visual impact (favoring things currently being displayed) but still give an opportunity to match images not being displayed due to bad metadata.
It “may” be worth exploring reducing the time interval for the case with invalid homography, image not displayed, as an invalid homography indicates the object is likely visible.